+++ web/src/components/studies/StudySelection.tsx
@@ -1,29 +1,14 @@
 import Button from "@/components/ui/Button";
-import InfoTooltip from "@/components/ui/InfoTooltip";
 import { Study } from "@/openapi";
 import { useRouter } from "next/router";
 
 import styles from "./StudySelection.module.css";
+import StudyStatusBadge from "../ui/StudyStatusBadge";
 
 type StudySelectionProps = {
   studies: Study[];
   isAdmin: boolean;
 };
-
-function getStatusDescription(status: string | undefined): string {
-  switch (status) {
-    case "Incomplete":
-      return "Click 'manage study' to complete your study setup.";
-    case "Pending":
-      return "Study is under review and awaiting approval from administrators.";
-    case "Approved":
-      return "Study has been approved and is ready for use.";
-    case "Rejected":
-      return "Study has been rejected. Please review feedback and make necessary changes.";
-    default:
-      return "Status information not available.";
-  }
-}
 
 export default function StudySelection(props: StudySelectionProps) {
   const { studies, isAdmin } = props;
@@ -37,12 +22,7 @@ export default function StudySelection(props: StudySelectionProps) {
         {studies.map((study) => (
           <div key={study.id} className={styles["study-card"]}>
             <div className={styles["status-indicator"]}>
-              <span className={`${styles["status-badge"]} ${styles[`status-${study.approval_status?.toLowerCase()}`]}`}>
-                {study.approval_status}
-                <span className={styles["tooltip-wrapper"]}>
-                  <InfoTooltip text={getStatusDescription(study.approval_status)} />
-                </span>
-              </span>
+              <StudyStatusBadge status={study.approval_status} isAdmin={isAdmin} />
             </div>
 
             <div className={styles["study-info"]}>
